+++ pumpkin/src/plugin/api/events/player/unregister_channel.rs
